Move service to modules
